Address some PR feedbacks: rename some classes and revert unnecessary whitespace changed.
