+++ core/src/main/java/org/opensearch/sql/exception/UnsupportedCursorRequestException.java
@@ -1,0 +1,9 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.opensearch.sql.exception;
+
+public class UnsupportedCursorRequestException extends RuntimeException {
+}
