+++ core/src/main/java/org/opensearch/sql/executor/QueryService.java
@@ -70,7 +70,6 @@ public class QueryService {
     }
   }
 
-
   /**
    * Explain the query in {@link UnresolvedPlan} using {@link ResponseListener} to
    * get and format explain response.
