--- core/src/main/java/org/opensearch/sql/legacy/plugin/UnsupportCursorRequestException.java
@@ -1,9 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package org.opensearch.sql.legacy.plugin;
-
-public class UnsupportCursorRequestException extends RuntimeException {
-}
