+++ core/src/main/java/org/opensearch/sql/planner/physical/PaginateOperator.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.sql.planner;
+package org.opensearch.sql.planner.physical;
 
 import java.util.List;
 import lombok.EqualsAndHashCode;
