+++ core/src/main/java/org/opensearch/sql/planner/physical/PhysicalPlanNodeVisitor.java
@@ -6,7 +6,6 @@
 
 package org.opensearch.sql.planner.physical;
 
-import org.opensearch.sql.planner.PaginateOperator;
 import org.opensearch.sql.storage.TableScanOperator;
 import org.opensearch.sql.storage.write.TableWriteOperator;
