+++ core/src/test/java/org/opensearch/sql/planner/physical/PaginateOperatorTest.java
@@ -26,7 +26,6 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import org.opensearch.sql.expression.DSL;
-import org.opensearch.sql.planner.PaginateOperator;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 public class PaginateOperatorTest extends PhysicalPlanTestBase {
